+++ Engine/Source/Runtime/RenderCore/Private/RenderGraphUtils.cpp
@@ -414,7 +414,7 @@ void AddClearUAVPass(FRDGBuilder& GraphBuilder, FRDGTextureUAVRef TextureUAV, fl
 
 void AddClearUAVPass(FRDGBuilder& GraphBuilder, FRDGTextureUAVRef TextureUAV, const FVector& Value)
 {
-	AddClearUAVPass(GraphBuilder, TextureUAV, { Value.X, Value.Y , Value.Z , 0.f });
+	AddClearUAVPass(GraphBuilder, TextureUAV, { (float)Value.X, (float)Value.Y , (float)Value.Z , 0.f });	// LWC_TODO: Precision loss?
 }
 
 void AddClearUAVPass(FRDGBuilder& GraphBuilder, FRDGTextureUAVRef TextureUAV, const FIntPoint& Value)
